Feature/basic auth

+++ backend/app/Http/Controllers/Api/UserController.php
@@ -3,11 +3,14 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\LoginUserRequest;
 use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
 use App\Http\Resources\UserResource;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -16,6 +19,7 @@ class UserController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', User::class);
         return UserResource::collection(User::all());
     }
 
@@ -25,6 +29,8 @@ class UserController extends Controller
     public function store(StoreUserRequest $request)
     {
         $user = User::create($request->validated());
+
+        $user->assignRole($request->role);
 
         return new UserResource($user);
     }
@@ -42,7 +48,11 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+        Gate::authorize('update', $user);
+
         $user->update($request->validated());
+
+        $user->syncRoles([$request->role]);
 
         return new UserResource($user);
     }
@@ -52,8 +62,51 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        Gate::authorize('delete', $user);
+
         $user->delete();
 
         return response()->noContent();
+    }
+
+    public function login(LoginUserRequest $request) {
+        if (!Auth::attempt($request->only('email', 'password'))) {
+            return response()->json(['message' => 'Invalid credentials'], 401);
+        }
+
+        $request->session()->regenerate();
+
+        return response()->json([
+            'user' => $request->user(),
+        ]);
+    }
+
+    public function me(Request $request) {
+        return new UserResource($request->user());
+    }
+
+    public function logout(Request $request)
+    {
+        Auth::logout();
+
+        $request->session()->invalidate();
+        $request->session()->regenerateToken();
+
+        return response()->json(['message' => 'Logged out successfully.'], 200);
+    }
+
+    public function loginWithToken(LoginUserRequest $request) {
+        if (!Auth::attempt($request->only('email', 'password'))) {
+            return response()->json(['message' => 'Invalid credentials'], 401);
+        }
+
+        $user = $request->user();
+        $token = $user->createToken('api-token');
+
+        return ['token' => $token->plainTextToken];
+    }
+
+    public function meWithToken(Request $request) {
+        return new UserResource($request->user());
     }
 }
